Fix hypre test shell scripting sadness
The pipefile rubbish makes configure fail on bash for no reason.

Now it gets far enough to whinge about a few missing wrappers in
cusparse, which is trivial for when someone gets bored.

+++ hypre/01-build.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
-set -ETeuo pipefail
 SCRIPT_DIR="$(realpath "$(dirname "$0")")"
 source "${SCRIPT_DIR}"/../util/args.sh "$@"
+
+MPI_DIR="${OUT_DIR}/openmpi/install"
+if [ ! -e "${MPI_DIR}" ] ; then
+    echo "Please build the OpenMPI third party project first. Use the same working directory." 1>&2
+    exit 1
+fi
 
 # Clean up any previous stuff.
 rm -rf "${OUT_DIR}/hypre/build"
@@ -13,16 +18,16 @@ cp -r "${OUT_DIR}/hypre/hypre" "${OUT_DIR}/hypre/build"
 # Configure
 SM="$(echo "${GPU_ARCH}" | sed -E 's/sm_//g')"
 
-export NVCC_PREPEND_FLAGS="-ccbin ${CUDA_PATH}/bin"
-
 cd "${OUT_DIR}/hypre/build/src"
-HYPRE_CUDA_SM="${SM}" \
-    ./configure --with-cuda --with-gpu-arch="${SM}" --disable-onemklsparse --disable-onemklblas --disable-onemklrand
+./configure \
+  --enable-unified-memory \
+  --with-cuda \
+  --with-gpu-arch="${SM}" \
+  --disable-onemklsparse \
+  --disable-onemklblas \
+  --disable-onemklrand \
+  --with-MPI-include="${MPI_DIR}/include" \
+  --with-MPI-libs="mpi" \
+  --with-MPI-lib-dirs="${MPI_DIR}/lib" \
 
-# Build.
-if [ "${VERBOSE}" == "1" ] ; then
-    VERBOSE="VERBOSE=1"
-else
-    VERBOSE=
-fi
-make test -j"${BUILD_JOBS}" "${VERBOSE}"
+make test -j$(nproc)
